Move getSerialization to utils.js
